Create first controller trigger from POST message
This organizes the code so that our Bot gets started immediately upon
load. This likely will not scale, since it is a single RTM for all
different messages.

We need to store slack_user_ids on our DB. Then we can set CRON jobs on
the DB. When the time triggers, it can find the appropriate slack
userID on the DB, then send back a POST API that will ping the right
user appropriately